LayoutTests/fast/js/kde/function_length-expected.txt

PASS Array.prototype.toString.length is 0
PASS Array.prototype.toLocaleString.length is 0
PASS Array.prototype.concat.length is 1
PASS Array.prototype.join.length is 1
PASS Array.prototype.pop.length is 0
PASS Array.prototype.push.length is 1
PASS Array.prototype.reverse.length is 0
PASS Array.prototype.shift.length is 0
PASS Array.prototype.slice.length is 2
PASS Array.prototype.sort.length is 1
PASS Array.prototype.splice.length is 2
PASS Array.prototype.unshift.length is 1
PASS Boolean.prototype.toString.length is 0
PASS Boolean.prototype.valueOf.length is 0
PASS Date.prototype.toString.length is 0
PASS Date.prototype.toUTCString.length is 0
PASS Date.prototype.toDateString.length is 0
PASS Date.prototype.toTimeString.length is 0
PASS Date.prototype.toLocaleString.length is 0
PASS Date.prototype.toLocaleDateString.length is 0
PASS Date.prototype.toLocaleTimeString.length is 0
PASS Date.prototype.valueOf.length is 0
PASS Date.prototype.getTime.length is 0
PASS Date.prototype.getFullYear.length is 0
PASS Date.prototype.getUTCFullYear.length is 0
PASS Date.prototype.toGMTString.length is 0
PASS Date.prototype.getMonth.length is 0
PASS Date.prototype.getUTCMonth.length is 0
PASS Date.prototype.getDate.length is 0
PASS Date.prototype.getUTCDate.length is 0
PASS Date.prototype.getDay.length is 0
PASS Date.prototype.getUTCDay.length is 0
PASS Date.prototype.getHours.length is 0
PASS Date.prototype.getUTCHours.length is 0
PASS Date.prototype.getMinutes.length is 0
PASS Date.prototype.getUTCMinutes.length is 0
PASS Date.prototype.getSeconds.length is 0
PASS Date.prototype.getUTCSeconds.length is 0
PASS Date.prototype.getMilliseconds.length is 0
PASS Date.prototype.getUTCMilliseconds.length is 0
PASS Date.prototype.getTimezoneOffset.length is 0
PASS Date.prototype.setTime.length is 1
PASS Date.prototype.setMilliseconds.length is 1
PASS Date.prototype.setUTCMilliseconds.length is 1
PASS Date.prototype.setSeconds.length is 2
PASS Date.prototype.setUTCSeconds.length is 2
PASS Date.prototype.setMinutes.length is 3
PASS Date.prototype.setUTCMinutes.length is 3
PASS Date.prototype.setHours.length is 4
PASS Date.prototype.setUTCHours.length is 4
PASS Date.prototype.setDate.length is 1
PASS Date.prototype.setUTCDate.length is 1
PASS Date.prototype.setMonth.length is 2
PASS Date.prototype.setUTCMonth.length is 2
PASS Date.prototype.setFullYear.length is 3
PASS Date.prototype.setUTCFullYear.length is 3
PASS Date.prototype.setYear.length is 1
PASS Date.prototype.getYear.length is 0
PASS Date.prototype.toGMTString.length is 0
PASS Error.prototype.toString.length is 0
PASS eval.length is 1
PASS parseInt.length is 2
PASS parseFloat.length is 1
PASS isNaN.length is 1
PASS isFinite.length is 1
PASS escape.length is 1
PASS unescape.length is 1
PASS Math.abs.length is 1
PASS Math.acos.length is 1
PASS Math.asin.length is 1
PASS Math.atan.length is 1
PASS Math.atan2.length is 2
PASS Math.ceil.length is 1
PASS Math.cos.length is 1
PASS Math.exp.length is 1
PASS Math.floor.length is 1
PASS Math.log.length is 1
PASS Math.max.length is 2
PASS Math.min.length is 2
PASS Math.pow.length is 2
PASS Math.random.length is 0
PASS Math.round.length is 1
PASS Math.sin.length is 1
PASS Math.sqrt.length is 1
PASS Math.tan.length is 1
PASS Object.prototype.toString.length is 0
PASS Object.prototype.valueOf.length is 0
PASS RegExp.prototype.exec.length is 0
PASS RegExp.prototype.test.length is 0
PASS RegExp.prototype.toString.length is 0
PASS String.fromCharCode.length is 1
PASS String.prototype.concat.length is 1
PASS String.prototype.toString.length is 0
PASS String.prototype.valueOf.length is 0
PASS String.prototype.charAt.length is 1
PASS String.prototype.charCodeAt.length is 1
PASS String.prototype.indexOf.length is 1
PASS String.prototype.lastIndexOf.length is 1
PASS String.prototype.match.length is 1
PASS String.prototype.replace.length is 2
PASS String.prototype.search.length is 1
PASS String.prototype.slice.length is 2
PASS String.prototype.split.length is 2
PASS String.prototype.substr.length is 2
PASS String.prototype.substring.length is 2
PASS String.prototype.toLowerCase.length is 0
PASS String.prototype.toUpperCase.length is 0
PASS String.prototype.big.length is 0
PASS String.prototype.small.length is 0
PASS String.prototype.blink.length is 0
PASS String.prototype.bold.length is 0
PASS String.prototype.fixed.length is 0
PASS String.prototype.italics.length is 0
PASS String.prototype.strike.length is 0
PASS String.prototype.sub.length is 0
PASS String.prototype.sup.length is 0
PASS String.prototype.fontcolor.length is 1
PASS String.prototype.fontsize.length is 1
PASS String.prototype.anchor.length is 1
PASS String.prototype.link.length is 1
PASS Number.prototype.toString.length is 1
PASS Number.prototype.valueOf.length is 0
PASS Number.prototype.toFixed.length is 1
PASS successfullyParsed is true
